+++ crates/sdk-core/tests/integ_tests/workflow_tests/determinism.rs
@@ -434,7 +434,7 @@ impl ChildWfIdOrTypeChangeWf {
                 UntypedWorkflow::new("child"),
                 RawValue::new(vec![]),
                 ChildWorkflowOptions {
-                    workflow_id: "I'm bad and wrong!".to_string(),
+                    workflow_id: Some("I'm bad and wrong!".to_string()),
                     ..Default::default()
                 },
             )
@@ -444,7 +444,7 @@ impl ChildWfIdOrTypeChangeWf {
                 UntypedWorkflow::new("not the child wf type"),
                 RawValue::new(vec![]),
                 ChildWorkflowOptions {
-                    workflow_id: "1".to_string(),
+                    workflow_id: Some("1".to_string()),
                     ..Default::default()
                 },
             )